Merging using //eve/development/MAINLINE<-//eve/sandbox/RNB-BENCHMARK
Changed priority of local files under the res folder versus files in stuff files or remote cache (downloaded on demand). In a non-packaged client (as used in development) local files now have priority. This allows developers to rely on download-on-demand, with local overrides. A developer working on static data, for example, can have the static data loaded from files on the hard drive and the rest of the resources be downloaded.

This is implemented by wrapping the file access methods in a file system class, with an IBlueResFileSystem interface. There are now three implementations of this file system, local, stuff and remote. The local one is initialized in C++ on startup - the other two are registered in Python in autoexec, with the priority depending on blue.pyos.packaged.

The behavior of the packaged client is not changed.

This integration also includes minor fixes to EVE Probe and its build process.

Binaries submitted in 850213
Build time : 17 minutes 9.930000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/blue/": change = 850193]