Fix sys.path creation in Python interpreter mode https://jira.ccpgames.com/browse/EVEDEV-10440
Our modified stackless interpreter does not follow the standard Python logic for constructing sys.path. As such any PYTHONPATH passed in from the outside was not correctly obeyed, and that would break, for example, using our exefile's interpreter mode as PyCharm's Python console.
With this change we're at least making sure that the environmental PYTHONPATH is considered when constructing sys.path in interpreter mode. But really, the sys.path customization for exefile should simply be done without requiring this "unicode import" non-sense we built into our stackless fork. For example, a reasonable way of doing this could be to simply clear out sys.path when not running in interpreter mode before we add our lib directives to it.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/blue/": change = 1793272]