Fix an issue where a force disconnect freezes the macOS client
When the macOS client received a forceful disconnect we'd see massive memory increase. This is because StacklessIO behaves slightly different than CarbonIO in such an event. It appears that in the event of a forceful disconnect we would receive many empty packets, which StacklessIO still used to increment the packet sequence number. This in turn lead to machonet expecting packages that would never arrive. In turn, the packet queue in the client would fill up indefinitely, eating up as much memory as it could from the machine before eventually crashing.

With this change StacklessIO no longer increases the packet sequence number in case of those empty packets. This then yields a `(None, None, 0)` response tuple to SocketGPS which we unfortunately need to deal with in machonet. On the plus side, this make the macOS client behave a lot better than the Windows client in the event of a forced disconnect: There's only one pop up to offer a client restart or quit option.
Additionally, we are also putting an upper bound onto the packet queue to stop the scenario of missed packages eating up all memory. It will instead now raise a RuntimeError exception and cause a client to consider itself as disconnected.

For testing: be advised that this is an incredibly low level change to our networking code. Please watch out for increased reports of disconnects during the next mass test.

https://jira.ccpgames.com/browse/EO-12427

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/blue/": change = 1941519]