Fix resource leak in blue's YamlReader on macOS
Re-enabling bluetests and running them on macOS highlighted that blue's YamlReader would leak IRoot instances under certain failure conditions. Therefore we are now adding an extra `Unlock()` call to ensure that no reference to the created instance is kept around.

I am not entirely sure why this was a macOS only problem; presumably we're relying on implementation defined C++ behaviour and clang is therefore generating a different behaviour than MSVC. Trying to reproduce or proof such differences with a minimum compilable example was not successful so far, which is unsurprising given YamlReader's complexity.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/blue/": change = 2089889]