Update in-client memory graph viewer to display only Python and Total memory graphs and Expose explicit structure for memory and cpuUsage from blue
Exposing an explicit structure for blue.pyos.cpuUsage greatly improves its readability and usage. Furthermore, it no longer provides the "blue memory" and "other memory" values. Tracking of these values was incorrect and thus always required additional debugging steps for developers anyway. Therefore the usefulness of these values (and graphs) was limited, as they did not reliably tell us whether it was our code or an SDK that was leaking memory.
As part of exposing the explicit structure another issue came to light: What is now "taskletsProcessed" was sometimes interpreted as the number of tasklets processed, but it really was always the number of tasklets queued before the Python tick happened. Now it is always the number of tasklets processed during the tick. However, it is easy to retrieve the previous number as well, if so desired: simply add taskletsProcessed and taskletsQueued.

https://jira.ccpgames.com/browse/EO-33406
https://jira.ccpgames.com/browse/EO-33407

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/blue/": change = 2102069]