Add CPU frequency to blue.sysinfo.cpu
There are various places in our Python codebase which incorrectly determine the CPU's frequency using the `GetCycles` function. That function returns the frequency at which the performance counter runs, not the CPU. At the same time, there was no obvious way to ask for this information properly. Therefore a new attribute - `frequency` - is added to `blue.sysinfo.cpu` which is sourced from the PDM and returns the correct CPU frequency in MHz.

https://jira.ccpgames.com/browse/EO-35099

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/blue/": change = 2129751]