Fix serializing PyUnicode_Object
The old serialization of strings and unicode objects was a bit all over the place after the initial conversion to Python 3's C-API. On top of that it was using some C APIs marked as deprecated in Python 3. All of this is now fixed by removing the assumption-based UTF8 / UTF16 specific handling and instead writing out the unicode kind.
String table support is still broken, though.