Fix `WriteStream`'s buffer protocol implementation
A wrong length was returned when requesting buffer information for `WriteStream` instances: the `mSize` member tells us how much memory was allocated for the `WriteStream`, but we're really only interested in the amount of data written to memory, which is indicated by `mPos`.
This wrong length resulted in `Bogus map data in marshal stream` errors when storing `TY_REFERENCE` data, because the reference map is stored at the end of the stream. And the data there was rubbish when `mSize` != `mPos`.