Invert `blue.net` dependency on Python code
Turns out that, upon initialization, `blue.net` wants to import the `carbon.common.lib.const` module to receive the values for `ADDRESS_TYPE_CLIENT` and `ADDRESS_TYPE_NODE`. This is going to be problematic for the Python 3 branch, so we're inverting this dependency and expect the Python code to forward these values using newly exposed methods.