Add AddressType enum with values from Python
One could argue that duplicating these enum values in C++ is brittle, which it certainly is. However, given these constants never change in practice, and are coming from the platform layer anyway. Thus, in my opinion, the benefits of duplicating the values in C++ far outweigh the potential downside of this approach.

If anything, this should be considered "going forward" with moving more aspects of machoNet into C++.