Add `blue` to `sys.modules` upon initialization
If we don't do this then we get double initialization upon the first discovered `import blue` statement, which causes a heap of interesting issues. For example, it breaks the build flavour mechanism, and it throws a few interesting assertions from our C++ singletons.
Therefore, we treat `blue` as an imported module to begin with.