Fix crash when verifying manifest file
Python 3 changes the way in which the internal threading state is set up in the interpreter, e.g. this now happens implicitly when calling `Py_Initialize`. However, the manifest file wants to be verified before that call happens, and opening the manifest file relies on the resource file system ... which in turn wants to allow threading when doing any file I/O. But this crashed now because we cannot call `PyEval_SaveThread` before `Py_Initialize()` has returned successfully.