Fix TaskletTimer initialization
TaskletTimer instances would initialize their `PyObject*` members as part of their constructor. Generally this is a Good Thing.
However, in the specific case of `BluePyOS::mTTimer` this is a problem, because this means that we're calling Python C API functions before the C API has been initialized, and that's playing with fire. Surprisingly we got away with this in Python 2 already, even though this was problematic there as well.
Therefore, the Python object initialization is now moved to an explicit call that needs to happen later.