Add missing imports of internal modules
Python 3 cannot import those from Python source code for some reason, presumably we're missing a customization to the interpreter's import mechanism. In any case, since we're bundling multiple C extensions in the same binary, and because we initialized them in C++ land already, it's a no-brainer to simply add them to `sys.modules` during initialization.