Re-enable support for `sys.getpymalloced()`
This used to be a customization of the stackless python interpreter, most certainly done that way because the memory allocators required an interpreter modification. However, now that we can pass the allocators in from the outside, there is no more need for this to be an interpreter customization. Instead, we're simply adding it to the `sys` module when initializing `blue`, just like we do with the `clockthis` functionality.

It's always going to be 0 right now, however, because the memory allocators are not yet hooked up.

https://ccpgames.atlassian.net/browse/PLAT-2066