Accept bytes and unicode strings in `blue.PythonEvents.write()`
Our `PythonEvents` class that overrides `stdout` lacks a `buffer` attribute. That causes some Python packages - most notably `teamcity_messages` - to write bytes directly to `stdout`, which was so far not supported. It's probably more correct to expose a `buffer` attribute, but this is quicker to implement. At one point the `PythonEvents` approach needs to be revisited any way since it's lacking even more functionality compared to Python's file-like objects.