Add ability to run `bluetests` via `-m`
This is required for tests which depend on our customized socket module (e.g. anything involving network functionality). The `unittest` module imports `socket` and thus we can no longer monkeypatch it. With this functionality we have a chance to monkeypatch the socket module in order to run the tests. Furthermore, we can run `unittest.main()` on a tasklet, allowing tests to behave transparently asynchronous, like the rest of the Python codebase.
The implementation is largely inspired by `unittest`'s own `__main__.py` module.