Fix marshalling of unicode strings
Fun times! So the test suite was passing because `assertBlueObjectsEqual()` only asserted value equality for `blue.BlueWrapper` instances. Otherwise, type identity was enough. This masked an issue with saving / loading unicode strings, where the assumption was that it would be "good enough" to (de-)serialize the kind of the object and its underlying buffer directly. This was initially done to save the cost of enforcing a uniform encoding (e.g. UTF-8). This turned out to be a wrong assumption.

So now we convert to UTF-8 explicitly, which is relatively expensive on the CPU side, but shaves off some bytes of the payload because textual data will now always transmit in the shortest possible byte sequence.

Additionally, an explicit check for string value identity in `assertBlueObjectsEqual` was added to ensure those scenarios do not slip through the cracks any longer. A general value comparison yielded more test failures, so this is not going to be introduced in this commit but as some follow up work.