Improve Marshal unit tests
This adds a bunch of missing comparison operators because they are required to check for equality instead of the default identity check. This should probably also be done for the standard `blue.BlueWrappedType` type used by `blue-exposure`, as that would remove the need for having the special `assertBlueObjectsEqual` helper method.
Furthermore, there are no more "old-school" objects in Python 3, so the tests had become meaningless.