Fix flaky tests for process/thread times (#21)
https://ccpgames.atlassian.net/browse/PLAT-3476 (Flaky unittests regarding process and thread times)

The problem here is that the tests assume that systemTime and userTime are only 0 when the C++ code fails to get the process/thread times. However, CcpGetProcessTimes and CcpGetThreadTimes can return 0 after successfully retrieving these values from the OS specific API, if the API function is called early enough in the process and the time spent between periodic timer updates on the OS is sufficiently large, such as on Windows where a 16ms update interval for such timers seems to be normal.

Raising an OSError  when failing to get process / thread times.