Vendor `_ssl` Python extension
A vendorized copy of Python's `_ssl` extension is required because it relies on `PySocketModule_APIObject`. Without vendoring, `_ssl` would refer to PYthon's standard `_socket` module. This leads to crashes because in `blue`'s environment the standard `_socket` module is replaced with a stackless flavoured version that interacts with the `CarbonIO` / `StacklessIO` technologies.

Keeping this as a separate binary means that all standard Python import rules will keep working.