Ensure the mainScheduler exists in game and interpreter mode
Currently, it's recommended to have a scheduler reference around before calling some of the scheduler API. This was already done in game mode, but now it needs to be available in interpreter mode, too. So we simply move the call to the blue module initialization, instead of the game mode runner method.