Use `CCP_MSIZE` to track memory amount allocated by Python
The OS is free to allocate more than the requested amount of bytes, for example to keep a certain alignment of the underlying memory. As a consequence, simply adding the requested allocation size to the memory allocation counter means that it can underflow when using `CCP_MSIZE` to subtract the size of the allocated memory during the `free` operation.
Furthermore, `CCP_MSIZE` may return a negative value on some platforms to indicate failure (f.e. when asking for size of a pointer allocated on a different heap), in which case no subtraction can happen. This is not a problem when allocating memory because there it is known that the pointer is good and from the same heap.