Forward branch path to C++ Unit tests (#8)
This environment variable is present during discovery, but is not guaranteed to exist when the C++ unit tests are run. This is mildly inconvenient since when running through CLion the environment variable must be set both for the CMake build and for the run configuration for the tests.

Explicitly setting the ENVIRONMENT property when discovering tests, so that the variable is guaranteed to be set when the tests are run, without the need for additional configuration.