Remove C++ import of scheduler
This was a work-around so that build flavors could work for `carbon-scheduler`, and to accommodate for the fact that `carbon-scheduler` wasn't following the naming conventions. Special casing created a bunch of undesired side-effects, for example, the `carbon-io` module (and thus `socket`) cannot be imported before `blue` has been imported. Now that `carbon-scheduler` has been fixed up then the import done here is no longer necessary.