Yield instead of sleeping
Since our IO event loop requires us to tick it from the thread it lives on, we can't afford to waste time sleeping. Instead, yield in order to allow other threads to run, then allow the OS to switch back to our thread as soon as it wants to.

This prompts the removal of NextScheduledEvent in the BeOS interface, since we no longer need to keep track of when the next scheduled event is in order to decide how long we want to sleep for.

Calling NextScheduledEvent( 0 ) used to prevent sleep, but this change does not provide a replacement to stop the thread from yielding. This is because it's simpler and yielding every once in a while is healthy, even if the thread is busy.