Bring idle frame rate up to ~100 frames per second
Due to oversleeping, the actual frame rate when idling was closer to 90 frames per second than 100. Changing the sleep amount so that we aim for 1 millisecond before the desired frame rate, thus expecting some oversleep and bringing the frame rate in my test code up to ~100 frames per second.

Also preventing sleeps if the sleep would last less than one millisecond, preferring to yield instead. This is intended to save the processor the overhead of switching in and out of the thread.