Fix `TracyZone` `C4251` compiler warning
This warning originates from a public header file.
Any external component which includes `BlueStatistics.h` will
also emit this warning when building.
This causes issues for components such as `Trinity`
which treat build warnings as errors.

The reason for the warning is that the implementation of
`std:optional` is dependent on the compiler. Exporting a
symbol of that type through a DLL is therefore generally
unsafe as there can be no __general__ guarantee that any
two DLLs were built using the same toolchain.

With that said, this is warning is not indicative of a problem
for Carbon components where the toolchain is pre-determined.
The warning only becomes an issue when elevated
in severity.