Set Tracy program name
This way discovering the correct process to profile can be somewhat easier, depending on the passed in name.
Re-using the existing but now obsolete "server or dump file name" parameter for this to keep the amount of changes from the outside to be low. Should get a "proper" clean-up in the Python 3 version since we can break some backwards compatibility there.