Fix client locking up when stopping profiler
The problem here was that simply destroying the profiler using `tracy::ShutdownProfiler` causes the tracy network thread to wait indefinately for more communication. Calling `tracy::RequestShutdown` first stops the network communication thread, and thus avoids the lock up.
However, there still is an outstanding issue that requires terminating the session in the tracy visualizer tool in order to restart profiling in the same exefile process session.