Fix Tracy invalid zone order instrumentation failure
`TracyCZone` start and end order need to mimic stack behaviour.
Keys passed to the `g_taskletZoneStore` map through `TracyEnterZone`
and  `TracyLeaveZone` need to be unique, a side effect of being a map,
and also accessible from the instrumentation code.

Passing `self` as a key to `g_taskletZoneStore` is problematic because
`self` is a pointer to the module in-memory where the relevant function
is exposed from. `self` is therefore not a reliable, unique key.

This result in some skipped zone creations. Regardless a given zone is
added, the instrumentation code always calls `TracyLeaveZone` on
the same key, resulting in an out-of-order instrumentation failure
as Tracy is expecting a different zone to end.

Using a frame object fixes the issue as the object is unique to the
current frame. This, however, imposes a requirement towards
instrumented python code where  `TracyLeaveZone` needs to be called
from the same stack frame as where the call to `TracyEnterZone` is made.

Reference:
https://ccpgames.atlassian.net/browse/PLAT-8399