Fix crash when calling `StopTelemetry` without `StartTelemetry` being called first
Managing the state of the connection to a telemetry solution using multiple booleans is not very maintainable. Case in point, with the recent change to request the tracy profiler to shut down, a crash condition was introduced when calling `StopTelemetry` without ever calling `StartTelemetry`. This is because Tracy expects that profiler has been initialized before it's requested to shut down, and if it wasn't initialized then bad memory is accessed.
Alas, this simplifies a lot of the nested if-else logic around the booleans by introducing a state enumeration. While this is still not perfect, then at least checking for the correct state is a lot simpler than before, and therefore it's more straight forward to correctly transition between states, as well as use the expected update behaviour for the current state.

https://ccpgames.atlassian.net/browse/PLAT-8456