Delay Tracy instrumentation until profiler is connected
This change fixes an issue where the Tracy profiler would emit
a zone double end failure if connected to early at the start of the
client lifetime.

Ending a zone prior to connecting to the profiler should be a no-op as
Tracy is guarded against such eventualities via the Telemetry context
active state. This does not appear to sufficiently guard against such
eventualities as ending a zone on an empty zone stack.