Stop tracy profiler from fully shutting down
An edge case exists in Tracy where long-living threads exceeding
the lifetime of a single profiler session delimited by
`tracy::StartupProfiler` and `Tracy::ShutdownProfiler` will cause a
crash. This means we cannot fully shutdown the profiler in a safe manner
 without putting the application at risk.

The workaround provided in this commit removes the ability for the tracy
 client to be fully disabled via `tracy::ShutdownProfiler` when calling
 `blue.statistics.StopTelemetry`. The consequence of this is that it's
 not longer possible to fully eliminate profiling overhead after a
 profiling session is started.

See https://github.com/wolfpld/tracy/issues/999 for further details.