Use stack container as tasklet zone store
The key map used to record zones had two drawbacks. Firstly, it was not
possible to determine the last zone entered for a given fiber. This is
required by some of the API, `PyAppendToZone` as an example.

Secondly, and edge case exists where instrumentation may attempt to add
multiple zones sharing a frame object to the tasklet zone store.
Since the frame object is used as a key to the underlying map, we could
only have one zone per frame object active at any moment in time,
resulting in some missing instrumentation.

Another implementation detail is moving to the tasklet store to thread
local storage. This is the recommended approach acc. to the
documentation, and there are good reasons for it. Most importantly,
non-fiber zones should never intermingle across threads as it's not safe
to leave a zone from a thread which did not enter it.