Delay fiber name erasure following tasklet closure
The profiler may not receive valid fiber name string information for
short-lived tasklets if the underlying fiber name is deallocated
immediately upon tasklet closure.

This change delays fiber name deallocation by five seconds. This is
a temporary measure as it's not a great synchronization mechanism.
Ideally we'd be able to clear the fiber name store when the profiler is
disconnected, but that'll require us to mark all reported zones as
inactive, which we don't support at the time of writing.

Reference:
https://ccpgames.atlassian.net/browse/PLAT-8855