Source `openssl` using vcpkg
A version override for `openssl` is specified for two reasons:
* The baseline version of `openssl` is incomparable with the
versioning scheme for `1.1.1k`. This causes `vcpkg` minimal version
selection following all constraints placed at the top level to fail. The
reason for this is that `vcpkg` cannot compare the minimal version
requested and the baseline version.
* `carbon-io` has not been built against later versions of `openssl`,
I've confirmed it cannot link against the baseline version. We therefore
need to pin the `openssl` version to `1.1.1k`.

Unfortunately this will also need to be overridden for other components
consuming `carbon-blue` as a dependency, both direct and transitive,
until `carbon-io` has been built against a more recent version of
`openssl`.