Fix `Py_buffer` handling in `Marshal::Load`
The `Py_buffer` instance was not correctly released after switching from the old buffer protocol to the new one. This can cause memory leaks because each `Py_buffer` instance holds a strong reference to the object for which the buffer was requested. Additionally, the usage of `Py_DECREF` may have introduced silent memory corruptions. It's interesting that this didn't cause larger issues earlier.