Add curl dependency to vcpkg manifest
This is the closest available version to version `7.76.1` which is
available from the monolith. Curl is not part of blue's interface
and it's statically linked. In theory this should allow blue to
source a version of curl which is not ABI compatibly with the version
of curl in the monolith.