Adapt unit tests to vcpkg installation
Most of the complexity removed in this change involved copying around
libraries around in order for  them to be co-located as runtime
dependencies when running tests. This is made much more simple due to
the fact that vcpkg installs dependencies in a co-located manner. As an
example, `_carbonsocket.pyd` and `CcpCore.dll` are always installed in
the same location, removing the need to copy `CcpCore.dll` to be
co-located with `_carbonsocket.pyd`. The rest