Fix zlib-ng linkage
The zlib package contained in the monolith is zlib-ng, not zlib.
Additionally, the monolith exposes the zlib-ng package target as
ZLIB::ZLIB. This conflicts with the target exposed by the zlib port,
which is an indirect dependency of blue. This change ensures we link
carbon-blue against zlib-ng. Additionally, zlib and zlib-ng were made
statically linked.