Make `blue.DBRow`'s virtual columns read-only
The virtual column mechanism in `blue.DBRow` is a convenient way to avoid creating custom types that wrap a result returned from a DB call. Unfortunately, it is also a convenient way for a malicious actor to smuggle malicious payloads over the network. Specifically, using a virtual column's setter method can be used to protect attributes on spoofed machoNet packets from being overridden by certain sanity checks. By making the virtual columns read-only it is possible to detect if this is about to happen.

Ideally, the whole virtual column mechanism would be removed, because the getter of a virtual column can still be any callable that is whitelisted in `blue.marshal`. Unfortunately, virtual columns and `blue.DBRow` are used all over the place in the Python code, making it virtually impossible to remove this mechanism without rewriting large parts of the game code. However, these virtual columns are only used to read values - they are never set. Therefore, this limits the ways in which executable code can be injected.