Add check to DBRow to ensure that unmarshalling creates expected DBRowDescriptorCheck
When unmarshalling a DBRow the DBRowDescriptor type is read from the marshal string.
It is possible to change this value to create a different object.
However, the object must be in the whitelist for it to be created.
It is very likely that changing the object type created would result in a crash,
but for safety a check is added.

It is also a good indicator that the input marshal string has been manually altered.