Configure macOS unit tests to run sequentially by default
Running unit tests in parallel on macOS has been observed
to be unreliable across a number of test suites.

Unit tests were not run in parallel prior to sourcing Blue's
Kotlin DSL in-source. This is therefore not to be considered
a regression in the build pipeline.

This is a temporary solution pending a proper investigation
outlined in the ticket below:

https://ccpgames.atlassian.net/browse/PLAT-9857