Save Unicode strings in Python 2.7 compatible format
The strings were saved as straight-up Unicode data, rather than as a byte string containing escape characters, which is what the 2.x version expects. In order to enable running a mixed-mode cluster, save strings in the old format.

The backwards compatibility with 2.7 should be temporary. Therefore, add a compile definition in order to make it easy to identify and remove the relevant code, once support for this feature ends.