Remove `marshaltesthelper.py`
This file was used in order to provide a non-global namespace for classes
being unmarshalled. The reason being that attempting to un-marshal a class
under the global namespace will result in a runtime error being thrown for
security reason.

The new approach uses the same namespace as the symbols defined in `test_marshal.py`.