Fix `test_load_object_with_data` backwards compatibility test
A python 2.7 marshalled object will have all `__dict__` keys encoded as byte
string. These keys require a conversion on the python 3.12 side of the fence.

There is some overhead involved with this operation as we now construct two
state objects. This is deemed acceptable because there is no other elegant way to
overwrite dictionary keys while iterating over the same set of keys.