Handle Python2.7 string unmarshalling in a consistent manner
Marshalling in Python2.7 approaches most strings as byte objects.
This is mostly compatible with Python3.12 except in cases where
special-case strings are being marshalled such as single-character
strings.

In those cases, the approach undertaken by the python environment
differs between major versions. Single-character strings are marshalled
as string objects in Python 2.7, which do not comply with Python3.12
unicode strings. Since we're unable to determine which major version of
python a particular marshalled object stems from, we cannot decide
whether a single-character marshalled object should be treated as a
unicode object or a byte object.