Parse row names using `s#` format unit
`s*` returns a Py_buf structure. The tests are likely succeeding because
the first set of elements in the buffer represent the row name,
but we really should be explicit in what we're expecting returned from
the call to `PyArg_ParseTuple`.

The `s#` format unit returns a similar set of values to `s`, except the
`const char *` value returned represents a "borrowed buffer" according
to the documentation. It's not clear what python object this is in
reference to.