Fix memory leak when converting object keys to Unicode
`PyUnicode_FromEncodedObject` returns a new reference, which will need
to have its reference count decreased at some point. This happens when
`encodedKey` key falls out of scope. The call to `PyDict_SetItem`
increments the reference count of the key again before the key has its
reference count decreased when falling out of scope. When it falls out
of scope, the reference count will be equal to `1`.

Finally, `PyDict_Update` will decrease the reference count of all keys
belonging to the `other` object passed into the function call.
Therefore, the ownership of each `encodedKey` is transferred to the new
object.