Allow marshalled new-style objects to construct as old-style objects
Python3 marshals all class objects as new-style objects. This is a
problem because an old-style object marshalled as a new-style object may
not be constructable as a new-style object in Python 2. In those cases,
we cannot assume the class definition of the object in question contains
a `__new__` operator, and must therefore create the class instance
using a similar approach to old-style objects in
`Marshall::ReadObjectsInstance`.