Prevent noise when starting up Python interpreter from causing invalid tests
Since the string "test" is used for pattern matching by the unittest module when finding test cases, it is guaranteed to be a part of the path for any valid Python test. Filtering out strings that do not contain this seems to be sufficient to weed out the noise when starting up the interpreter batch file.